Displaying the details of an article
The article is displayed along with a form for the comment section
If the article's owner is connected, he can update it or publish | unpublish

Publishing or unpublishing the article is a matter of updating a property on the article table
After each of thos actions, we redirect into a specific page.
* On the list page we come from it
* On the article page itself otherwise

We ensure that after each update, the action is inserted into the history table
We also set a conditionnal statement that will put as viewed the current article that is displayed by the user
But if no one is connected, then the statement is not executed.

Along with the article infos, we have put the number of times that the article has been read (by identified users)
We do this by counting the number of accurence on the view table.