Changing the article behavior
We updated the display of the article by using `marked` to parse the content typed by the user.
We are also usiing `prisme` to display potential block of code that may be included into the article.

When adding a new article, we added a new input text to type the name of the article
Alse we displayed the same input in the update page that contain by default the article's categories.

Next we will focus on saving the article's categories