Adding a new feature : publishing a media along with the article
For that we used the known multer to get files from the user
On submit of the form, we check out to see if the file is valid (audio, video, image)

If so, we add it to the user's session (not directly to the database as the article is not created yet)

If the user choose to edit an article or to add one, we get all the data from the session an put it into the database associating it with either the newly created article or the existing one.

We offered also the user the ability to remove a media through a button.

Those requests (adding / removing  media) are performed asynchronously using the fetch API.