Refactoring the code
This time we haven't add any functionnalities but just refactored the code to reuse some building blocks like when the article is updated we have to do the same validation process as when the article is created.

Also, we centralized the access to the database so that it is opened at one place and accessed anywhere. And after each use of the database, we make sure to close the instance, that is (we think) a key to not waste ressources unecessarily.